Move test_verify_date() to test_download.py
